feat: Support integer suffix

+++ tests/primitives/native_cast/native_cast.c
@@ -3,11 +3,11 @@
 int milone_main(void);
 
 int milone_main(void) {
-    void* mutNull_ = ((void*)((uintptr_t)0));
+    void* mutNull_ = ((void*)(size_t)0ULL);
     void const* constNull_ = ((void const*)mutNull_);
     int const* constIntPtr_ = ((int const*)constNull_);
     int* intPtr_ = ((int*)constIntPtr_);
     intptr_t address_ = ((intptr_t)intPtr_);
-    milone_assert((((int)address_) == 0), 11, 2);
+    milone_assert((address_ == 0LL), 11, 2);
     return 0;
 }
